% Generated by roxygen2 (4.0.1): do not edit by hand
\name{draw.vis}
\alias{draw.vis}
\title{Interactively visualize summary trees in a browser}
\usage{
draw.vis(json.object, out.dir = tempfile(), open.browser = interactive())
}
\arguments{
\item{json.object}{character string output from \link{prepare.vis()}.}

\item{out.dir}{directory to store html/js/json files.}

\item{open.browser}{Should R open a browser? If yes, this function will
attempt to create a local file server via the servr package.
This is necessary since the javascript needs to access local files and most
browsers will not allow this.}
}
\value{
An invisible object.
}
\description{
Interactively visualize a set of summary trees in a browser
}
\details{
This function will place the necessary html/js/css files (located in
\code{system.file("vis", package = "summarytrees")}) in a directory specified
by \code{out.dir}, start a local file server in that directory (if necessary),
and (optionally) open the default browser in this directory.
}

