% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Gauss}
\alias{Gauss}
\title{Carl Gauss subtree of the Math Genealogy Tree}
\format{A data frame with 43527 rows and 4 variables:
\describe{
  \item{node}{node ID number for each mathematician, numbered from
              1 to n = 43527 by default (note that these are not the same ID
              numbers used by the Mathematics Genealogy Project)}
  \item{parent}{the node ID number of the parent of a given
                mathematician}
  \item{weight}{set to 1 by default for each mathematician}
  \item{label}{the full name of the mathematician}
}}
\source{
This data is distributed with the permission of and in cooperation
with the Mathematics Genealogy Project (MGP):
\url{http://genealogy.math.ndsu.nodak.edu/}.

Also note that:

1. This data may not be redistributed separately from the R package.

2. This is a snapshot of MGP data from June 7, 2012.

3. The node identifiers (1, 2, ..., 43537) here are not the same
identifiers used by the MGP (we mention this for clarification purposes).
}
\usage{
Gauss
}
\description{
A list of student-advisor relationships of mathematicians who descended from
Carl Gauss.
}
\keyword{datasets}

