% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.tree}
\alias{get.tree}
\title{Convert node-parent data to parent-children data}
\usage{
get.tree(nodes = integer(), parents = integer(), weights = numeric(),
  labels = character())
}
\arguments{
\item{nodes}{integer vector containing a set of positive, unique, integers
denoting the identities of the nodes of the tree}

\item{parents}{integer vector containing the (positive integer) id of the
parent of each node in the tree. One  and only one element of
\code{parents} must be set to zero, indicating which node is the root of
the tree.}

\item{weights}{numeric vector containing the nonnegative weight of
each node in the tree. Used in this function only for breaking ties
in ordering of children.}

\item{labels}{character vector containing the labels of
each node in the tree. Used in this function only for breaking ties
in ordering of children.}
}
\value{
tree an integer matrix with three columns and as many rows
as there are parent nodes in the tree.
The first column contains the indices of the parent nodes in
the tree. The second column contains the index of the first child
node of a given parent node, and the third column contains the index
of the last child of a given parent node.
}
\description{
Convert two integer vectors containint node identifiers and
their respective parent identifiers to a data structure that
contains a list of all parents, and for each parent, the first
and last child identifier, after re-naming all nodes to be
in decreasing order of level, and such that all children of
a given parent node are given consecutive integer identifiers.
This is a utility function that takes the user-friendly
node-parent data format and converts it to the parent-children
format that is currently required by the function that does the
computation of the maximum entropy summary trees.
}

