% Generated by roxygen2 (4.0.1): do not edit by hand
\name{greedy}
\alias{greedy}
\title{Compute summary trees using greedy algorithm}
\usage{
greedy(tree = matrix(), weights = numeric(), labels = character(),
  K = integer())
}
\arguments{
\item{structure}{an integer matrix with three columns and \code{P}
rows. The first column contains the indices of the parent nodes in
the tree. The second column contains the index of the first child
node of a given parent node, and the third column contains the index
of the last child of a given parent node.}

\item{weights}{numeric vector containing non-negative weight of each node
in the tree.}

\item{labels}{character vector containing labels for each node in the tree}

\item{K}{integer. The number of nodes in the largest summary tree.}
}
\value{
A length-K list of matrices, where each matrix represents the
summary tree with K nodes. The kth matrix contains k rows and 4 columns.
}
\description{
Compute a series of summary trees for a given
input tree using a greedy algorithm. This algorithm returns summary
trees whose entropies are typically within 5% of the maximum
possible entropy (based on a series of experiments). The
algorithm runs substantially faster than the optimal algorithm.
}
\references{
\url{http://www2.research.att.com/~kshirley/papers/KarloffShirleyWebsite.pdf}
}

