% Generated by roxygen2 (4.0.1): do not edit by hand
\name{optimal}
\alias{optimal}
\title{Compute maximum entropy summary trees using exact algorithm}
\usage{
optimal(node = integer(), parent = integer(), weight = numeric(),
  label = character(), K = integer(), epsilon = 0)
}
\arguments{
\item{node}{integer vector containing a set of positive, unique, integers
denoting the identities of the nodes of the tree}

\item{parent}{integer vector containing the (positive integer) id of the
parent of each node in the tree. One  and only one element of
\code{parents} must be set to zero, indicating which node is the root of
the tree.}

\item{weight}{numeric vector containing non-negative weight of each node
in the tree.}

\item{label}{character vector containing labels for each node in the tree}

\item{K}{integer. The number of nodes in the largest summary tree.}

\item{epsilon}{the nonnegative upper bound of the additive error when
using the approximate algorithm. Each k-node summary tree returned by
the approximate algorithm will have entropy within \code{epsilon} of the
maximal entropy of a k-node summary tree.}
}
\value{
A length-K list of matrices, where each matrix represents the
summary tree with K nodes. The kth matrix contains k rows and 4 columns.
(describe the columns later...)
}
\description{
Compute a series of summary trees for a given
input tree using an exact algorithm.
}
\references{
\url{http://www2.research.att.com/~kshirley/papers/KarloffShirleyWebsite.pdf}
}

