% Generated by roxygen2 (4.0.1): do not edit by hand
\name{order.nodes}
\alias{order.nodes}
\title{Order the nodes of the input tree}
\usage{
order.nodes(node = integer(), parent = integer(), weight = numeric(),
  label = character())
}
\arguments{
\item{node}{integer vector containing a set of positive, unique, integers
denoting the identities of the nodes of the tree}

\item{parent}{integer vector containing the (positive integer) id of the
parent of each node in the tree. One  and only one element of
\code{parents} must be set to zero, indicating which node is the root of
the tree.}

\item{weight}{numeric vector containing the nonnegative weight of
each node in the tree. Used in this function only for breaking ties
in ordering of children.}

\item{label}{character vector containing the labels of
each node in the tree. Used in this function only for breaking ties
in ordering of children.}
}
\value{
list containing two elements: \code{tree}, an integer matrix
with three columns and as many rows
as there are parent nodes in the tree.
The first column contains the indices of the parent nodes in
the tree. The second column contains the indices of the first child
node of each parent node, and the third column contains the indices
of the last child node of each parent node. Second, \code{data}, a
data.frame containing five columns: node, parent, weights, labels,
and level, which is equal to 1 for the root of the tree. This
data.frame is returned because the ordering of the nodes may be
changed by running this function.
}
\description{
Convert two integer vectors containing node identifiers and
their respective parent identifiers to a data structure that
contains a list of all parents, and for each parent, the first
and last child identifier, after re-ordering all nodes to be
in decreasing order of level, and such that all children of
a given parent node have consecutive integer identifiers.
This is a utility function that takes the user-friendly
node-parent data format and converts it to the parent-children
format that is currently required by the function that does the
computation of the maximum entropy summary trees.
}

