% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse.optimal.stdout}
\alias{parse.optimal.stdout}
\title{Parse the stdout of the optimal algorithm C program}
\usage{
parse.optimal.stdout(stdout = character())
}
\arguments{
\item{stdout}{character vector containing, in each element, one line of
standard output from the C implementation of the optimal algorithm.}
}
\value{
A length-K list of matrices, where each matrix represents the
summary tree with K nodes. The kth matrix contains k rows and 3 columns:
(1) the node number, (2) the node weight, and (3) the type of node in the
summary tree, which is one of 3 types: 1 = singleton, 2 = subtree, or
3 = 'other' node. If node type = 3, the node number is the parent
node of this 'other' cluster.
}
\description{
This function parses the text that is printed to standard output
after the C program to implement the optimal algorithm is run.
Eventually we'll replace this, but for now it's the simplest
way to write an R wrapper around the C function originally
written by H. Karloff.
}
\references{
\url{http://www2.research.att.com/~kshirley/papers/KarloffShirleyWebsite.pdf}
}

