% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepare.vis}
\alias{prepare.vis}
\title{Prepare data to supply the interactive visualization}
\usage{
prepare.vis(tree.list, labels = NULL, tree = NULL, legend.width = 100,
  node.width = 150, node.height = 12, units = "", print.weights = FALSE,
  legend.color = "fed976", color.level = NULL)
}
\arguments{
\item{tree.list}{the list of summary trees that is output by either the
\code{greedy()} function or the \code{optimal()} function.}

\item{labels}{a character vector of labels for all the nodes in the
original tree (to be depracated).}

\item{tree}{a 3-column matrix representing the tree as output by either
the \code{greedy()} function or the \code{optimal()} function (to be
depracated).}
}
\value{
json.object a JSON object that will be passed to the browser to
feed the interactive visualization
}
\description{
Prepare data from summary trees and create one json object
to supply necessary information for the browser to create interactive
visualization.
}
\details{
This function takes the output from one of the functions that
computes summary trees -- \code{greedy()} or \code{optimal()} -- and
processes the output
to create the necessary file to feed the interactive visualiation.
}

