% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarytrees_htmlwidget.R
\name{renderSummarytrees_htmlwidget}
\alias{renderSummarytrees_htmlwidget}
\title{Widget render function for use in Shiny}
\usage{
renderSummarytrees_htmlwidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

