% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Rsource}
\alias{Rsource}
\title{File structure of the R software package}
\format{A data frame with 4704 rows and 4 variables:
\describe{
  \item{node}{node ID number for each file or directory, numbered from
              1 to n = 4704 by default}
  \item{parent}{the node ID number of the parent of each node}
  \item{weight}{the size of the file in bytes, where zero indicates
                a directory rather than a file}
  \item{label}{the name of the file or directory}
}}
\source{
\url{https://cran.r-project.org/}
}
\usage{
Rsource
}
\description{
A list of the file names and sizes in the R software package (version 3.2.1,
aka 'World Famous Astronaut', released 2015-06-18), including the
hierarchical structure of the directory.
}
\keyword{datasets}

