% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarytrees_htmlwidget.R
\name{summarytrees_htmlwidget}
\alias{summarytrees_htmlwidget}
\title{Interactively visualize summary trees in a browser}
\usage{
summarytrees_htmlwidget(json.object, width = NULL, height = NULL)
}
\arguments{
\item{json.object}{character string output from \link{prepare.vis}.}
}
\value{
An \code{htmlwidget} object.
}
\description{
Visualize your summary trees as an interactive \code{d3.js} using
               \code{htmlwidgets}.  The widget will work in various
               R contexts, such as RStudio viewer, R console, a web browser, or Shiny.
}
\examples{
library(summarytrees)

json <- paste0(
  readLines("http://www2.research.att.com/~kshirley/jsm2015/data.json")
  ,collapse="\\n"
)

summarytrees_htmlwidget(json)
}

