% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarytrees_htmlwidget.R
\name{summarytrees_htmlwidgetOutput}
\alias{summarytrees_htmlwidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
summarytrees_htmlwidgetOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

