% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytrees_htmlwidget.R
\name{summarytrees_htmlwidget}
\alias{summarytrees_htmlwidget}
\title{Interactively visualize summary trees in a browser}
\usage{
summarytrees_htmlwidget(json.object, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{json.object}{character string output from \link{prepare.vis}.}
}
\value{
An \code{htmlwidget} object.

}
\description{
Visualize your summary trees as an interactive \code{d3.js} using
               \code{htmlwidgets}.  The widget will work in various
               R contexts, such as RStudio viewer, R console, a web browser, or Shiny.

}
\examples{
library(summarytrees)
data(dmoz)

#use example from vignette
K <- 100
g <- greedy(node = dmoz[, "node"],
           parent = dmoz[, "parent"],
           weight = dmoz[, "weight"],
           label = dmoz[, "label"],
           K = K)

# Prepare the summary trees for visualization:
json <- prepare.vis(tree.list = g$summary.trees,
                    labels = g$data[, "label"],
                    tree = g$tree,
                    legend.width = 150,
                    node.width = 225,
                    node.height = 14,
                    units = "# of URLs",
                    print.weights = TRUE,
                    legend.color = "lightsteelblue",
                    color.level = 3)
summarytrees_htmlwidget(json)


}

